pyproject.toml
[project]
name = "btt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bokeh>=3.7.3",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-bootstrap>=3.3.7.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-moment>=1.0.6",
    "flask-paranoid>=0.3.0",
    "flask-script>=2.0.6",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "nltk>=3.9.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "py-dateutil>=2.2",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
]
